// app-microservices/microservices/src/backend/model/actionPerformedDetailsAssetsUnManaged.ts
/**
 * TIBCO DISCOVER backend microservice
 * Api Layer for the backend of Project Discover
 *
 * The version of the OpenAPI document: 1.0
 * Contact: fcenedes@tibco.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { UnManageDataSetInfoStored } from './unManageDataSetInfoStored';

export class ActionPerformedDetailsAssetsUnManaged {
    'message': string;
    'code': number;
    'tdv': UnManageDataSetInfoStored;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "message",
            "baseName": "message",
            "type": "string"
        },
        {
            "name": "code",
            "baseName": "code",
            "type": "number"
        },
        {
            "name": "tdv",
            "baseName": "tdv",
            "type": "UnManageDataSetInfoStored"
        }    ];

    static getAttributeTypeMap() {
        return ActionPerformedDetailsAssetsUnManaged.attributeTypeMap;
    }
}