// app-microservices/microservices/src/backend/model/actionPerformedLoginValidate.ts
/**
 * TIBCO DISCOVER backend microservice
 * Api Layer for the backend of Project Discover
 *
 * The version of the OpenAPI document: 1.0
 * Contact: fcenedes@tibco.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class ActionPerformedLoginValidate {
    'status': string;
    'code': number;
    'orgId': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "status",
            "baseName": "status",
            "type": "string"
        },
        {
            "name": "code",
            "baseName": "code",
            "type": "number"
        },
        {
            "name": "orgId",
            "baseName": "orgId",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return ActionPerformedLoginValidate.attributeTypeMap;
    }
}