// app-microservices/microservices/src/backend/model/previewConfigFile.ts
/**
 * TIBCO DISCOVER backend microservice
 * Api Layer for the backend of Project Discover
 *
 * The version of the OpenAPI document: 1.0
 * Contact: fcenedes@tibco.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class PreviewConfigFile {
    'Token': string;
    'Organization': string;
    'DatasetId': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "Token",
            "baseName": "Token",
            "type": "string"
        },
        {
            "name": "Organization",
            "baseName": "Organization",
            "type": "string"
        },
        {
            "name": "DatasetId",
            "baseName": "DatasetId",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return PreviewConfigFile.attributeTypeMap;
    }
}