// app-microservices/microservices/src/backend/model/redisFileInfo.ts
/**
 * TIBCO DISCOVER backend microservice
 * Api Layer for the backend of Project Discover
 *
 * The version of the OpenAPI document: 1.0
 * Contact: fcenedes@tibco.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class RedisFileInfo {
    'ContentType': string;
    'LastModified': string;
    'OriginalFilename': string;
    'OriginalEncoding': string;
    'FileSize': string;
    'newline': string;
    'EscapeChar': string;
    'QuoteChar': string;
    'Separator': string;
    'Encoding': string;
    'OriginalNewLine': string;
    'FileLocation': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "ContentType",
            "baseName": "ContentType",
            "type": "string"
        },
        {
            "name": "LastModified",
            "baseName": "LastModified",
            "type": "string"
        },
        {
            "name": "OriginalFilename",
            "baseName": "OriginalFilename",
            "type": "string"
        },
        {
            "name": "OriginalEncoding",
            "baseName": "OriginalEncoding",
            "type": "string"
        },
        {
            "name": "FileSize",
            "baseName": "FileSize",
            "type": "string"
        },
        {
            "name": "newline",
            "baseName": "newline",
            "type": "string"
        },
        {
            "name": "EscapeChar",
            "baseName": "EscapeChar",
            "type": "string"
        },
        {
            "name": "QuoteChar",
            "baseName": "QuoteChar",
            "type": "string"
        },
        {
            "name": "Separator",
            "baseName": "Separator",
            "type": "string"
        },
        {
            "name": "Encoding",
            "baseName": "Encoding",
            "type": "string"
        },
        {
            "name": "OriginalNewLine",
            "baseName": "OriginalNewLine",
            "type": "string"
        },
        {
            "name": "FileLocation",
            "baseName": "FileLocation",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return RedisFileInfo.attributeTypeMap;
    }
}